refactor(kernel): extract federation syscalls to kernel/src/kernel/federation.rs
Plan §G called for 8 syscall-family submodules (io, ipc, locks,
dispatch, observability, mount, federation, agents). This commit fills
in the federation submodule alongside the existing six.

Methods extracted from kernel/mod.rs into kernel/federation.rs:
  - set_distributed_coordinator / distributed_coordinator
  - zones_procfs_stat / zones_procfs_readdir
  - stash_blob_fetcher_slot / take_pending_blob_fetcher_slot

Methods stay on `Kernel` via `impl Kernel { ... }` blocks — same API
shape, just file-organization.

Agents intentionally not extracted: the agent table is a primitive in
`rust/kernel/src/core/agents/`, not a syscall family. The §6.1 dep
diagram covers this separately.

cargo check + clippy + fmt + codegen clean.
